basic test for replace_below_confidence
There is no public API to directly change confidence value,
so this is ideally not a true test. Having said that it
could have been done with a mix of using other methods.
Since, those methods are in another moule like mmath_aggr,
so I dont want to create its dependency here.